Move components into more lowercasey folder structures.
