+++ src/components/SEO/index.jsx
@@ -1,7 +1,7 @@
 import React, { Component } from "react";
 import Helmet from "react-helmet";
 import urljoin from "url-join";
-import config from "../../../data/SiteConfig";
+import config from "../../../data/site-config";
 
 class SEO extends Component {
   render() {
