+++ src/pages/about/index.jsx
@@ -1,8 +1,8 @@
 import React, { Component } from "react";
 import Helmet from "react-helmet";
-import Layout from "../layout";
-import About from "../components/About/About";
-import config from "../../data/SiteConfig";
+import Layout from "../../layout";
+import About from "../../components/about";
+import config from "../../../data/site-config";
 
 class AboutPage extends Component {
   render() {
