--- src/templates/post.css
@@ -1,5 +1,0 @@
-.post-meta {
-    display: flex;
-    flex-direction: column;
-    justify-content: center;
-}
